preliminar_7

+++ src/model/entities/active/Character.java
@@ -7,8 +7,8 @@ public class Character extends Active
 
 	private String playerName;
 
-	public Character (	String id, Point pos, double ps, double vel, double maxVel,
-						Point offset, String playerName )
+	public Character (	String id, Point pos, int ps, double vel, double maxVel, Point offset,
+						String playerName )
 	{
 		super(id, pos, ps, vel, maxVel, offset);
 		this.playerName = playerName;
